Move repeatable alarm to Fragment
